+++ ArcdpsLogManager/Configuration/Settings.cs
@@ -1,4 +1,5 @@
 using GW2Scratch.ArcdpsLogManager.Configuration.Migrations;
+using GW2Scratch.ArcdpsLogManager.Sections.Clears;
 using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;
@@ -40,6 +41,8 @@ namespace GW2Scratch.ArcdpsLogManager.Configuration
 			MinimumLogDurationSecondsChanged += (sender, args) => SaveToFile();
 			IgnoredUpdateVersionsChanged += (sender, args) => SaveToFile();
 			DpsReportUploadDetailedWvwChanged += (sender, args) => SaveToFile();
+			PlayerAccountNamesChanged += (sender, args) => SaveToFile();
+			WeeklyClearGroupsChanged += (sender, args) => SaveToFile();
 
 			return LoadFromFile();
 		});
@@ -276,6 +279,32 @@ namespace GW2Scratch.ArcdpsLogManager.Configuration
 				}
 			}
 		}
+		
+		public static IReadOnlyList<string> PlayerAccountNames
+		{
+			get => Values.PlayerAccountNames;
+			set
+			{
+				if (!Equals(Values.PlayerAccountNames, value))
+				{
+					Values.PlayerAccountNames = value.ToList();
+					OnPlayerAccountNamesChanged();
+				}
+			}
+		}
+		
+		public static IReadOnlyList<EncounterCategory> WeeklyClearGroups
+		{
+			get => Values.WeeklyClearGroups;
+			set
+			{
+				if (!Equals(Values.WeeklyClearGroups, value))
+				{
+					Values.WeeklyClearGroups = value.ToList();
+					OnWeeklyClearGroupsChanged();
+				}
+			}
+		}
 
 		public static event EventHandler<EventArgs> LogRootPathChanged;
 		public static event EventHandler<EventArgs> ShowDebugDataChanged;
@@ -290,6 +319,8 @@ namespace GW2Scratch.ArcdpsLogManager.Configuration
 		public static event EventHandler<EventArgs> MinimumLogDurationSecondsChanged;
 		public static event EventHandler<EventArgs> IgnoredUpdateVersionsChanged;
 		public static event EventHandler<EventArgs> DpsReportUploadDetailedWvwChanged;
+		public static event EventHandler<EventArgs> PlayerAccountNamesChanged;
+		public static event EventHandler<EventArgs> WeeklyClearGroupsChanged;
 
 		private static void OnLogRootPathsChanged()
 		{
@@ -354,6 +385,16 @@ namespace GW2Scratch.ArcdpsLogManager.Configuration
 		private static void OnDpsReportUploadDetailedWvwChanged()
 		{
 			DpsReportUploadDetailedWvwChanged?.Invoke(null, EventArgs.Empty);
+		}
+		
+		private static void OnPlayerAccountNamesChanged()
+		{
+			PlayerAccountNamesChanged?.Invoke(null, EventArgs.Empty);
+		}
+		
+		private static void OnWeeklyClearGroupsChanged()
+		{
+			WeeklyClearGroupsChanged?.Invoke(null, EventArgs.Empty);
 		}
 	}
 }