+++ ArcdpsLogManager/Configuration/StoredSettings.cs
@@ -1,8 +1,7 @@
+using GW2Scratch.ArcdpsLogManager.Sections.Clears;
 using GW2Scratch.ArcdpsLogManager.Uploads;
-using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;
-using System.IO;
 
 namespace GW2Scratch.ArcdpsLogManager.Configuration;
 
@@ -25,4 +24,7 @@ public class StoredSettings
 	public int? MinimumLogDurationSeconds { get; set; } = null;
 	public List<string> HiddenLogListColumns { get; set; } = new List<string>() {"Character", "Map ID", "Game Version", "arcdps Version", "Instabilities", "Scale"};
 	public List<string> IgnoredUpdateVersions { get; set; } = new List<string>();
+	public List<string> PlayerAccountNames { get; set; } = new List<string>();
+	public List<EncounterCategory> WeeklyClearGroups { get; set; } =
+		[EncounterCategory.Raids, EncounterCategory.StrikeEndOfDragons, EncounterCategory.StrikeSecretsOfTheObscure];
 }