+++ ArcdpsLogManager/Sections/Clears/EncounterAvailability.cs
@@ -1,0 +1,8 @@
+namespace GW2Scratch.ArcdpsLogManager.Sections.Clears;
+
+public enum EncounterAvailability
+{
+	Available,
+	DoesNotExist,
+	NotLogged,
+}