+++ ArcdpsLogManager/Sections/Clears/EncounterCategory.cs
@@ -1,0 +1,12 @@
+using System.Collections.Generic;
+
+namespace GW2Scratch.ArcdpsLogManager.Sections.Clears;
+
+public enum EncounterCategory
+{
+	// Do not change the backing values of these enums, they are used in stored settings.
+	Raids = 0,
+	StrikeIcebroodSaga = 1,
+	StrikeEndOfDragons = 2,
+	StrikeSecretsOfTheObscure = 3,
+}